+++ src/newold/handlers/JarHandlerImplementation.ts
