+++ src/newold/handlers/TemplateHandlerImplementation.ts
