+++ src/newold/interfaces/ExceptionHandler.ts
