+++ src/newold/interfaces/TemplateHandler.ts
